requirements.txt
coverage==5.5
curlylint==0.13
djangorestframework==3.12.4
# covered by djangorestframework: Django
flake8==3.8.4
flake8-django==1.1.2
gunicorn==20.0.4
numpy==1.21
psycopg2-binary==2.9.1
python-dotenv==0.18
social-auth-app-django==5.0.0
graphviz==0.17
django-extensions==3.1.3
pydotplus==2.0.2